requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
channels==4.3.0
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==3.0.1
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
phonenumbers==9.0.10
priority==2.0.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
